Add ZONE_AWARE
