+++ pom.xml
@@ -198,7 +198,7 @@
               com.hazelcast.spi.discovery,com.hazelcast.core,
               com.hazelcast.logging,com.hazelcast.nio, com.hazelcast.util,
               com.hazelcast.internal.json,javax.net.ssl,javax.security.auth.x500,
-              javax.naming, javax.naming.directory
+              javax.naming, javax.naming.directory, com.hazelcast.spi.partitiongroup
             </Import-Package>
           </instructions>
         </configuration>
